feat: add plan details section
