+++ ui/src/pages/organizations/details/side-panel/index.tsx
@@ -1,14 +1,8 @@
-import { Avatar, Flex, List, Text, SidePanel } from "@raystack/apsara/v1";
-import styles from "./side-panel.module.css";
+import { Avatar, SidePanel } from "@raystack/apsara/v1";
 import { V1Beta1Organization } from "~/api/frontier";
-import dayjs from "dayjs";
-import { useEffect, useState } from "react";
-import { api } from "~/api";
-import Skeleton from "react-loading-skeleton";
-import { V1Beta1Subscription } from "@raystack/frontier/api-client";
-import { V1Beta1Plan } from "@raystack/frontier";
 import { OrganizationDetailsSection } from "./org-details-section";
 import { KYCDetailsSection } from "./kyc-section";
+import { PlanDetailsSection } from "./plan-details-section";
 
 export const SUBSCRIPTION_STATES = {
   active: "Active",
@@ -17,64 +11,6 @@ export const SUBSCRIPTION_STATES = {
   canceled: "Canceled",
   "": "NA",
 } as const;
-
-const BillingDetails = ({ organizationId }: { organizationId: string }) => {
-  const [subscription, setSubscription] = useState<V1Beta1Subscription>();
-  const [plan, setPlan] = useState<V1Beta1Plan>();
-
-  useEffect(() => {
-    async function fetchBillingDetails(id: string) {
-      const subResponse = await api?.frontierServiceListSubscriptions2(id);
-      const subscriptions = subResponse?.data?.subscriptions || [];
-      const sub = subscriptions.find(
-        (sub) => sub.state === "active" || sub.state === "trialing",
-      );
-      if (sub && sub.plan_id) {
-        setSubscription(sub);
-        const planResponse = await api?.frontierServiceGetPlan(sub.plan_id);
-        const plan = planResponse?.data?.plan;
-        setPlan(plan);
-      }
-    }
-    if (organizationId) {
-      fetchBillingDetails(organizationId);
-    }
-  }, [organizationId]);
-
-  return (
-    <List.Root>
-      <List.Header>Plan details</List.Header>
-      <List.Item>
-        <List.Label className={styles["side-panel-section-item-label"]}>
-          Name
-        </List.Label>
-        <List.Value>{plan?.title || "N/A"}</List.Value>
-      </List.Item>
-      <List.Item>
-        <List.Label className={styles["side-panel-section-item-label"]}>
-          Started from
-        </List.Label>
-        <List.Value>
-          {dayjs(subscription?.current_period_start_at).format("DD MMM YYYY")}
-        </List.Value>
-      </List.Item>
-      <List.Item>
-        <List.Label className={styles["side-panel-section-item-label"]}>
-          Ends on
-        </List.Label>
-        <List.Value>
-          {dayjs(subscription?.current_period_end_at).format("DD MMM YYYY")}
-        </List.Value>
-      </List.Item>
-      <List.Item>
-        <List.Label className={styles["side-panel-section-item-label"]}>
-          Payment mode
-        </List.Label>
-        <List.Value>Prepaid</List.Value>
-      </List.Item>
-    </List.Root>
-  );
-};
 
 interface SidePanelProps {
   organization: V1Beta1Organization;
@@ -94,7 +30,7 @@ export function OrgSidePanel({ organization }: SidePanelProps) {
         <KYCDetailsSection organizationId={organization.id || ""} />
       </SidePanel.Section>
       <SidePanel.Section>
-        <BillingDetails organizationId={organization.id || ""} />
+        <PlanDetailsSection organizationId={organization.id || ""} />
       </SidePanel.Section>
     </SidePanel>
   );
